Modified

+++ src/ECS/Tilecomponent.hpp
@@ -16,9 +16,9 @@ public:
     {
         SDL_DestroyTexture(texture);
     }
-    TileComponent(int srcX, int srcY, int xpos, int ypos,int tsize,int tscale, std:: string id)
+    TileComponent(int srcX, int srcY, int xpos, int ypos, int tsize, int tscale, std::string id)
     {
-        texture = Game:: assets->GetTexture(id);
+        texture = Game::assets->GetTexture(id);
         position.x = xpos;
         position.y = ypos;
         srcRect.x = srcX;
@@ -26,12 +26,12 @@ public:
         srcRect.w = srcRect.h = tsize;
         destRect.x = xpos;
         destRect.y = ypos;
-        destRect.w = destRect.h = tsize*tscale;
+        destRect.w = destRect.h = tsize * tscale;
     }
     void update() override
     {
-        destRect.x = position.x - Game::camera.x;
-        destRect.y = position.y - Game::camera.y;
+        destRect.x = static_cast<int>(position.x - Game::camera.x);
+        destRect.y = static_cast<int>(position.y - Game::camera.y);
     }
     void draw() override
     {
